requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
alabaster==1.0.0
arrow==1.3.0
astroid==3.3.9
asttokens==3.0.0
asyncssh==2.20.0
atomicwrites==1.4.1
attrs==25.3.0
autopep8==2.0.4
babel==2.17.0
backports.tarfile==1.2.0
beautifulsoup4==4.13.3
binaryornot==0.4.4
black==25.1.0
bleach==6.2.0
Bottleneck==1.5.0
Cartopy==0.25.0
certifi==2025.1.31
cffi==1.17.1
cftime==1.6.4.post1
chardet==5.2.0
charset-normalizer==3.4.1
choreographer==1.0.10
clarabel==0.10.0
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
concave_hull==0.0.9
contourpy==1.3.1
cookiecutter==2.6.0
cryptography==44.0.2
cvxpy==1.6.4
cycler==0.12.1
dask==2025.5.1
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
deprecation==2.1.0
diff-match-patch==20241021
dill==0.3.9
docstring-to-markdown==0.15
docutils==0.21.2
ecos==2.0.14
executing==2.2.0
fastjsonschema==2.21.1
flake8==7.1.2
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.5.1
geopandas==1.1.1
highspy==1.11.0
idna==3.10
imagesize==1.4.1
importlib_metadata==8.6.1
inflection==0.5.1
intervaltree==3.1.0
ipykernel==6.29.5
ipython==8.34.0
isort==6.0.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
jellyfish==1.1.3
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kaleido==1.0.0
keyring==25.6.0
kiwisolver==1.4.8
linopy==0.5.5
locket==1.0.0
logistro==1.1.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
matplotlib-label-lines==0.8.0
mccabe==0.7.0
mdurl==0.1.2
mistune==3.1.2
more-itertools==10.6.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.30.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
netCDF4==1.7.2
networkx==3.4.2
numexpr==2.11.0
numpy==2.2.4
numpydoc==1.8.0
orjson==3.11.2
osqp==0.6.7.post3
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.2
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.1.0
platformdirs==4.3.6
plotly==6.3.0
pluggy==1.5.0
polars==1.31.0
prompt_toolkit==3.0.50
propcache==0.3.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycddlib==3.0.2
pycodestyle==2.12.1
pycountry==24.6.1
pycparser==2.22
pydocstyle==6.3.0
pyflakes==3.2.0
PyGithub==2.6.1
Pygments==2.19.1
PyJWT==2.10.1
pylint==3.3.5
pylint-venv==3.0.4
pyls-spyder==0.4.0
PyNaCl==1.5.0
pyogrio==0.11.0
pyparsing==3.2.1
pyproj==3.7.1
pypsa==0.35.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
PyQtWebEngine==5.15.7
PyQtWebEngine-Qt5==5.15.2
pyshp==2.3.1
python-dateutil==2.9.0.post0
python-lsp-black==2.0.0
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.12.2
python-slugify==8.0.4
pytoolconfig==1.3.1
pytz==2025.1
pyuca==1.2
pywin32==310
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq==26.3.0
QDarkStyle==3.2.3
qdldl==0.1.7.post5
qstylizer==0.2.4
QtAwesome==1.3.1
qtconsole==5.6.1
QtPy==2.4.3
referencing==0.36.2
requests==2.32.3
rich==13.9.4
roman-numerals-py==3.1.0
rope==1.13.0
rpds-py==0.23.1
rtree==1.4.0
scipy==1.15.2
scs==3.2.7.post2
seaborn==0.13.2
shapely==2.0.7
simplejson==3.20.1
six==1.17.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.6
Sphinx==8.2.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
spyder==6.0.4
spyder-kernels==3.0.3
stack-data==0.6.3
superqt==0.7.1
tabulate==0.9.0
text-unidecode==1.3
textdistance==4.6.3
three-merge==0.1.1
tinycss2==1.4.0
tomlkit==0.13.2
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
tqdm_joblib==0.0.5
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
urllib3==2.3.0
validators==0.35.0
watchdog==6.0.0
wcwidth==0.2.13
webencodings==0.5.1
whatthepatch==1.0.7
wrapt==1.17.2
xarray==2025.6.1
yapf==0.43.0
yarl==1.18.3
zipp==3.21.0
